refactor(webapp): add ProducersTable component and rename producers components
